Move new functions for RBAC v2 to rbac_v2.go
