+++ pilot/pkg/networking/plugin/authz/rbac.go
@@ -177,19 +177,6 @@ func (service serviceMetadata) match(rule *rbacproto.AccessRule) bool {
 	return true
 }
 
-// isServiceMatchedWorkload checks if the calling service is matched with the selector in the AuthorizationPolicy if
-// present.
-func (service serviceMetadata) isServiceMatchedWorkload(labels map[string]string) bool {
-	// Check if labels with |key| and |value| exist in service.
-	for key, value := range labels {
-		actualValue, present := service.labels[key]
-		if !present || actualValue != value {
-			return false
-		}
-	}
-	return true
-}
-
 // attributesFromAuthN returns true if the given attribute is generated from AuthN filter. This implies
 // the attribute should be enforced by dynamic metadata.
 func attributesFromAuthN(k string) bool {
@@ -354,10 +341,14 @@ func (Plugin) OnInboundListener(in *plugin.InputParams, mutable *plugin.MutableO
 		return nil
 	}
 	option := rbacOption{authzPolicies: authzPolicies, globalPermissiveMode: globalPermissive}
+	isRbacV2 := isRbacV2(attr.Namespace, option)
+	if isRbacV2 && isConflictingRbacVersion(attr.Namespace, option) {
+		return nil
+	}
 	switch in.ListenerProtocol {
 	case plugin.ListenerProtocolTCP:
 		rbacLog.Debugf("building tcp filter config for %v", *service)
-		filter := buildTCPFilter(service, option, util.IsProxyVersionGE11(in.Node))
+		filter := buildTCPFilter(service, option, util.IsProxyVersionGE11(in.Node), isRbacV2)
 		if filter != nil {
 			rbacLog.Infof("built tcp filter config for %s", service.name)
 			for cnum := range mutable.FilterChains {
@@ -366,7 +357,7 @@ func (Plugin) OnInboundListener(in *plugin.InputParams, mutable *plugin.MutableO
 		}
 	case plugin.ListenerProtocolHTTP:
 		rbacLog.Debugf("building http filter config for %v", *service)
-		filter := buildHTTPFilter(service, option, util.IsProxyVersionGE11(in.Node))
+		filter := buildHTTPFilter(service, option, util.IsProxyVersionGE11(in.Node), isRbacV2)
 		if filter != nil {
 			rbacLog.Infof("built http filter config for %s", service.name)
 			for cnum := range mutable.FilterChains {
@@ -438,11 +429,16 @@ func isRbacEnabled(svc string, ns string, authzPolicies *model.AuthorizationPoli
 	}
 }
 
-func buildTCPFilter(service *serviceMetadata, option rbacOption, is11 bool) *listener.Filter {
+func buildTCPFilter(service *serviceMetadata, option rbacOption, is11 bool, isRbacV2 bool) *listener.Filter {
 	option.forTCPFilter = true
 	// The result of convertRbacRulesToFilterConfig() is wrapped in a config for http filter, here we
 	// need to extract the generated rules and put in a config for network filter.
-	config := convertRbacRulesToFilterConfigV2(service, option)
+	var config *http_config.RBAC
+	if isRbacV2 {
+		config = convertRbacRulesToFilterConfigV2(service, option)
+	} else {
+		config = convertRbacRulesToFilterConfig(service, option)
+	}
 	tcpConfig := listener.Filter{
 		Name: rbacTCPFilterName,
 	}
@@ -464,9 +460,14 @@ func buildTCPFilter(service *serviceMetadata, option rbacOption, is11 bool) *lis
 
 // buildHTTPFilter builds the RBAC http filter that enforces the access control to the specified
 // service which is co-located with the sidecar proxy.
-func buildHTTPFilter(service *serviceMetadata, option rbacOption, is11 bool) *http_conn.HttpFilter {
+func buildHTTPFilter(service *serviceMetadata, option rbacOption, is11 bool, isRbacV2 bool) *http_conn.HttpFilter {
 	option.forTCPFilter = false
-	config := convertRbacRulesToFilterConfigV2(service, option)
+	var config *http_config.RBAC
+	if isRbacV2 {
+		config = convertRbacRulesToFilterConfigV2(service, option)
+	} else {
+		config = convertRbacRulesToFilterConfig(service, option)
+	}
 	rbacLog.Debugf("generated http filter config: %v", *config)
 	out := &http_conn.HttpFilter{
 		Name: rbacHTTPFilterName,
@@ -479,128 +480,6 @@ func buildHTTPFilter(service *serviceMetadata, option rbacOption, is11 bool) *ht
 	}
 
 	return out
-}
-
-// convertRbacRulesToFilterConfigV2 is the successor of convertRbacRulesToFilterConfig, which supports
-// converting AuthorizationPolicy.
-func convertRbacRulesToFilterConfigV2(service *serviceMetadata, option rbacOption) *http_config.RBAC {
-	rbac := &policyproto.RBAC{
-		Action:   policyproto.RBAC_ALLOW,
-		Policies: map[string]*policyproto.Policy{},
-	}
-	namespace := service.attributes[attrDestNamespace]
-	// Check to make sure we are not dereferencing null pointers.
-	if option.authzPolicies == nil || option.authzPolicies.NamespaceToAuthorizationConfigV2 == nil {
-		return &http_config.RBAC{Rules: rbac}
-	}
-	if _, present := option.authzPolicies.NamespaceToAuthorizationConfigV2[namespace]; !present {
-		return &http_config.RBAC{Rules: rbac}
-	}
-
-	// Get all AuthorizationPolicy Istio config from this namespace.
-	allAuthzPolicies := option.authzPolicies.NamespaceToAuthorizationConfigV2[namespace].AuthzPolicies
-	// If there is no AuthorizationPolicy in this namespace, try converting using the old converting function.
-	if len(allAuthzPolicies) == 0 {
-		return convertRbacRulesToFilterConfig(service, option)
-	} else if option.globalPermissiveMode {
-		// Permissive Mode will be ignored.
-		// TODO(pitlv2109): Handle permissive mode in the future.
-		return &http_config.RBAC{Rules: rbac}
-	}
-	// Get all key-value pairs of AuthorizationPolicy name to list of Subjects/ServiceRoleBindings in this namespace.
-	authzPolicyToSubjects := option.authzPolicies.AuthzPolicyToAllowSubjects(namespace)
-	for _, authzPolicy := range allAuthzPolicies {
-		// If a WorkloadSelector is used in the AuthorizationPolicy config and does not match this service, skip this
-		// AuthorizationPolicy.
-		if authzPolicy.Policy.WorkloadSelector != nil && !service.isServiceMatchedWorkload(authzPolicy.Policy.WorkloadSelector.Labels) {
-			continue
-		}
-		for i, subject := range authzPolicyToSubjects[authzPolicy.Name] {
-			// Convert role.
-			role := option.authzPolicies.GetServiceRoleFromName(namespace, subject.RoleRef.Name)
-			permission := &policyproto.Permission{}
-			rule := role.Rules[0]
-
-			// Check to make sure that the current service (caller) is the one this rule is applying to.
-			if !service.match(rule) {
-				continue
-			}
-			// Also check if we have converted this rule before.
-			if !hasConvertedPermission(rbac, subject.RoleRef.Name) {
-				rbacLog.Debugf("rules[%d] matched", i)
-				if option.forTCPFilter {
-					// TODO(yangminzhu): Move the validate logic to push context and add metrics.
-					if err := validateRuleForTCPFilter(rule); err != nil {
-						// It's a user misconfiguration if a HTTP rule is specified to a TCP service.
-						// For safety consideration, we ignore the whole rule which means no access is opened to
-						// the TCP service in this case.
-						rbacLog.Debugf("rules[%d] ignored, found HTTP only rule for a TCP service: %v", i, err)
-						continue
-					}
-				}
-				// Generate the policy if the service is matched and validated to the services specified in
-				// ServiceRole.
-				permission = convertToPermission(rule)
-				if permission.String() == "" {
-					rbacLog.Debugf("role %s skipped for no rule matched", subject.RoleRef.Name)
-					continue
-				}
-			}
-
-			if option.forTCPFilter {
-				if err := validateBindingsForTCPFilter([]*rbacproto.ServiceRoleBinding{subject}); err != nil {
-					rbacLog.Debugf("role %s skipped, found HTTP only binding for a TCP service: %v", subject.RoleRef.Name, err)
-					continue
-				}
-			}
-			// Convert binding.
-			enforcedPrincipals, _ := convertToPrincipals([]*rbacproto.ServiceRoleBinding{subject}, option.forTCPFilter)
-
-			if len(enforcedPrincipals) == 0 {
-				rbacLog.Debugf("role %s skipped for no principals found", subject.RoleRef.Name)
-				continue
-			}
-
-			if len(enforcedPrincipals) != 0 {
-				if rbac.Policies[subject.RoleRef.Name] == nil {
-					rbac.Policies[subject.RoleRef.Name] = &policyproto.Policy{
-						Permissions: []*policyproto.Permission{permission},
-						Principals:  enforcedPrincipals,
-					}
-				} else {
-					allPrincipals := rbac.Policies[subject.RoleRef.Name].Principals
-					// Check if enforcedPrincipals has existed in the allPrincipals.
-					if !isPrincipalInPrincipals(enforcedPrincipals, allPrincipals) {
-						rbac.Policies[subject.RoleRef.Name].Principals = append(allPrincipals, enforcedPrincipals...)
-					}
-				}
-			}
-		}
-	}
-	return &http_config.RBAC{Rules: rbac}
-}
-
-// isPrincipalInPrincipals returns True if the |principal| has already existed in the |principals|,
-// False otherwise.
-// This is to make sure we don't have duplicated principals for the same ServiceRole.
-// This can happen when we have a service foo that has more than one exact Subjects (in one or many
-// AuthorizationPolicy) that bind to the same ServiceRole.
-func isPrincipalInPrincipals(principal []*policyproto.Principal, principals []*policyproto.Principal) bool {
-	for _, currPrincipal := range principals {
-		if currPrincipal.String() == principal[0].String() {
-			return true
-		}
-	}
-	return false
-}
-
-// hasConvertedPermission returns True if there is already a |roleName| ServiceRole for |rbac|, False
-// otherwise.
-func hasConvertedPermission(rbac *policyproto.RBAC, roleName string) bool {
-	if rbac.Policies[roleName] == nil {
-		return false
-	}
-	return rbac.Policies[roleName].Permissions != nil
 }
 
 // convertRbacRulesToFilterConfig converts the current RBAC rules (ServiceRole and ServiceRoleBindings)
